requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-gis==0.14
pytz==2018.4
django-extensions==2.1.6
mysqlclient==1.3.13
pandas==0.24.2
django-datetime-widget==0.9.3
matplotlib==3.0.3
django-pandas==0.6.1
scipy==1.2.1
Pillow==6.0.0
openpyxl==2.6.2
